+++ CHANGELOG.md
@@ -4,6 +4,82 @@
 
 [1]: https://www.npmjs.com/package/gax-nodejs?activeTab=versions
 
+## [3.1.1](https://github.com/googleapis/gax-nodejs/compare/v3.1.0...v3.1.1) (2022-06-10)
+
+
+### Bug Fixes
+
+* update googlePRotoFilesDir ([#1280](https://github.com/googleapis/gax-nodejs/issues/1280)) ([e91f5ae](https://github.com/googleapis/gax-nodejs/commit/e91f5ae397868c742751fc537734e81d2293925b))
+
+## [3.1.0](https://github.com/googleapis/gax-nodejs/compare/v3.0.3...v3.1.0) (2022-06-07)
+
+
+### Features
+
+* support regapic LRO ([#1276](https://github.com/googleapis/gax-nodejs/issues/1276)) ([b9beedb](https://github.com/googleapis/gax-nodejs/commit/b9beedbc09ff18e326851dc5007c90d89b404561))
+
+### [3.0.3](https://github.com/googleapis/gax-nodejs/compare/v3.0.2...v3.0.3) (2022-05-26)
+
+
+### Bug Fixes
+
+* decode Http error using proto3-JSON-serializer ([#1273](https://github.com/googleapis/gax-nodejs/issues/1273)) ([17a790f](https://github.com/googleapis/gax-nodejs/commit/17a790f4d47bf588a6970052822f9fdb14acc321))
+* parse error from stream call ([#1275](https://github.com/googleapis/gax-nodejs/issues/1275)) ([141a587](https://github.com/googleapis/gax-nodejs/commit/141a58725c65ddd78003cbb94e56eb1d4f838067))
+
+### [3.0.2](https://github.com/googleapis/gax-nodejs/compare/v3.0.1...v3.0.2) (2022-05-23)
+
+
+### Bug Fixes
+
+* **deps:** update dependency proto3-json-serializer to v1 ([#1268](https://github.com/googleapis/gax-nodejs/issues/1268)) ([cff5c5c](https://github.com/googleapis/gax-nodejs/commit/cff5c5ce9dec1e43db3842e9cc44510a2c8abc0b))
+* **deps:** update dependency protobufjs to v6.11.3 ([#1272](https://github.com/googleapis/gax-nodejs/issues/1272)) ([6492a2c](https://github.com/googleapis/gax-nodejs/commit/6492a2cc55086439518dcb20cb6001c28e1d2bda))
+* **deps:** use protobufjs v6.11.3 ([#1271](https://github.com/googleapis/gax-nodejs/issues/1271)) ([d650c15](https://github.com/googleapis/gax-nodejs/commit/d650c1510cd98958fc32be03b890970f6f321595))
+
+### [3.0.1](https://github.com/googleapis/gax-nodejs/compare/v3.0.0...v3.0.1) (2022-05-11)
+
+
+### Bug Fixes
+
+* get protos from build/protos ([#1265](https://github.com/googleapis/gax-nodejs/issues/1265)) ([ea0d85c](https://github.com/googleapis/gax-nodejs/commit/ea0d85c0f25c734e62b5a11ef700cd81f04c587f))
+
+## [3.0.0](https://github.com/googleapis/gax-nodejs/compare/v2.30.3...v3.0.0) (2022-05-10)
+
+
+### ⚠ BREAKING CHANGES
+
+* remove pageSize/pageToken from call settings (#1263)
+* drop Node 10 to support and update lastest Typescript (#1252)
+* reduce module size by remove duplicate storage for protos files (#1249)
+
+### Bug Fixes
+
+* **deps:** update dependency retry-request to v5 ([#1264](https://github.com/googleapis/gax-nodejs/issues/1264)) ([26fbac8](https://github.com/googleapis/gax-nodejs/commit/26fbac8d178336532a12ac5cb484f0832f4cf2f8))
+* remove pageSize/pageToken from call settings ([#1263](https://github.com/googleapis/gax-nodejs/issues/1263)) ([fbf43d2](https://github.com/googleapis/gax-nodejs/commit/fbf43d2247c6b6bbbdb4c5962283d5dacb96c402))
+* revert gRPC stream change in [#1226](https://github.com/googleapis/gax-nodejs/issues/1226) ([#1257](https://github.com/googleapis/gax-nodejs/issues/1257)) ([1c3a5a7](https://github.com/googleapis/gax-nodejs/commit/1c3a5a7ac54f8fd8e710fe4316c3d5a82072f8f1))
+* upgrade proto-loader and add long dev dependency ([#1253](https://github.com/googleapis/gax-nodejs/issues/1253)) ([5dea136](https://github.com/googleapis/gax-nodejs/commit/5dea13698c580283083f41d330a1a2faf37adc5d))
+
+
+### Build System
+
+* drop Node 10 to support and update lastest Typescript ([#1252](https://github.com/googleapis/gax-nodejs/issues/1252)) ([88a180f](https://github.com/googleapis/gax-nodejs/commit/88a180f541a542fb6483e8353f8bd02582283b98))
+* reduce module size by remove duplicate storage for protos files ([#1249](https://github.com/googleapis/gax-nodejs/issues/1249)) ([4d2a04d](https://github.com/googleapis/gax-nodejs/commit/4d2a04d472ce1a6ebfc5494742d95d05ec726a81))
+
+### [2.30.3](https://github.com/googleapis/gax-nodejs/compare/v2.30.2...v2.30.3) (2022-05-03)
+
+
+### Bug Fixes
+
+* handle stream callback is undefined ([#1238](https://github.com/googleapis/gax-nodejs/issues/1238)) ([269f805](https://github.com/googleapis/gax-nodejs/commit/269f805e02039cfa537084f46a561fa527eba301))
+* isolate Rest Stream from retry logic to avoid backpressure ([#1226](https://github.com/googleapis/gax-nodejs/issues/1226)) ([485d5b2](https://github.com/googleapis/gax-nodejs/commit/485d5b2af79d389acf6cc613ce419830ccac6bef))
+* pin @grpc/proto-loader to v0.6.9 ([#1242](https://github.com/googleapis/gax-nodejs/issues/1242)) ([bca9a91](https://github.com/googleapis/gax-nodejs/commit/bca9a91f90795970d0bdc0c15ae8a3549d546eba))
+
+### [2.30.2](https://github.com/googleapis/gax-nodejs/compare/v2.30.1...v2.30.2) (2022-04-12)
+
+
+### Bug Fixes
+
+* **deps:** update dependency @grpc/grpc-js to ~1.6.0 ([#1207](https://github.com/googleapis/gax-nodejs/issues/1207)) ([2a16b76](https://github.com/googleapis/gax-nodejs/commit/2a16b76237094b42f1afb4667c82adebb87402b8))
+
 ### [2.30.1](https://github.com/googleapis/gax-nodejs/compare/v2.30.0...v2.30.1) (2022-03-08)